move mock implementation to class so we only need 1 shared context implementation
add README section about mock module

Note mock differences in README

README format

test mock module on iOS; add comments

NativeEventEmitter will always be defined in mock

run ios
